Extract sourcemap utilities into shared module
Add src/sourcemap-utils.js containing shared source-map helpers (getLineText, findNearestWord, offset/line conversions, mapToSourcePos, etc.). Remove duplicate implementations from src/typecheck.js and packages/vscode/src/lsp.js and update them to import/delegate to the new utilities. Also adjust gen->src fallback logic to walk mapped gen lines when resolving source lines. This centralizes sourcemap logic for both the CLI type-checker and the VS Code LSP and reduces duplicated code.
